+++ src/main/java/com/revature/revshop/model/Review.java
@@ -10,7 +10,7 @@ public class Review {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "review_id")
-    private Integer reviewId;
+    private Long reviewId;
 
     @ManyToOne
     @JoinColumn(name = "product_id", nullable = false)
@@ -40,11 +40,11 @@ public class Review {
         this.createdAt = LocalDateTime.now();
     }
 
-    public Integer getReviewId() {
+    public Long getReviewId() {
         return reviewId;
     }
 
-    public void setReviewId(Integer reviewId) {
+    public void setReviewId(Long reviewId) {
         this.reviewId = reviewId;
     }
